requirement.txt
alembic==1.4.2
BareNecessities==0.2.8
bcrypt==3.1.7
blinker==1.4
cffi==1.14.0
click==7.1.2
Faker==4.0.3
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-JWT-Extended==3.24.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.11.2
Mail==2.1.0
Mako==1.1.2
MarkupSafe==1.1.1
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-editor==1.0.4
six==1.14.0
slugify==0.0.1
SQLAlchemy==1.3.16
text-unidecode==1.3
Werkzeug==1.0.1
